requirements.txt
# 다중 경유지 최적화 프로그램 의존성

# CLI 인터페이스
click>=8.0.0

# 환경변수 관리
python-dotenv>=0.19.0

# Excel 파일 처리
pandas>=2.0.0
openpyxl>=3.0.0
xlrd>=2.0.0

# HTTP 요청 (카카오 API)
requests>=2.25.0

# 과학 계산 (호환성 고려)
numpy>=1.21.0,<2.0.0

# 머신러닝 (클러스터링, 선택적)
scikit-learn>=1.0.0

# 지도 시각화
folium>=0.12.0

# 개발/테스트용
pytest>=6.0.0
pytest-cov>=2.12.0